Prevent browser autocomplete of non-login password fields
User and password fields for channels and callbacks were being autocompleted with
the current user and pass. According to [1], browsers now ignore autocomplete off
in password fields, having to use new-password instead, to indicate that the
password is different from the site's.

Also, the password value was not being re-loaded, so any edit on the form (without
re-entering the password) ended up in the password value being cleared. Though
the correct way to fix this would be adding an option to "click here if you want
to change the password" and only render the control in that case, re-sending the
password back to the client does the trick for now.

Tested on Chrome 53, Firefox 47 and Safari 9.1.3

Fixes #717

[1]: https://developer.mozilla.org/en-US/docs/Web/Security/Securing_your_site/Turning_off_form_autocompletion

Conflicts:
	app/views/projects/_form.haml