Support filters by language and sms number in phonebook
Filters are supported in both phonebook and scheduled calls. All implicit vars are added to the variables list. See #713.

Ensure that the "value" being compared to is only sent to the server if the operator is not unary. Fixes #753.

Handle comparing two variables (either project or implicit) by equality. Fixes #754.