Add striping to some channel settings values
This should help a bit dealing with user entry errors while configuring
channels. Regarding #783, 2 out of 3 channels which fail to resolve the IP
address of the domain have a whitespace at the start/end of the domain, which
makes the DNS query fail. Plus, we've had authentication errors in the past due
to extra whitespace in the username and password.