Skip generating config for channels with unresolved domains
If the domain name of the channel cannot be resolved when generating the
configuration file for Asterisk, skip it entirely. Asterisk needs the resolved
IP addresses for matching incoming calls from the channel (ie. the `identify`
sections). Also, the domain needs to be resolvable for Asterisk to attempt the
registration.

We may consider relaxing this restriction if the channel should not be
registered and is only used for outgoing calls. As of today, the broker doesn't
care for the channel's direction.