Automatically disable SIP channels if registration fails repeatedly
Closes #782

On each check_status (every 30 seconds), increment a counter if the channel
registration failed. If this value exceeds a threshold, disable the channel and
rewrite the configuration so that the disabled channel is not written to the
config file. Counts are reset if the registration succeeds or the configuration
is re-written.

NB: Asterisk will stop registration attempts if it receives a 401/403 from the
registration server. A reload (executed every 2 minutes) will force it to try
again.