#824. Fix Outbound Calls in Africa's Talking Broker
The approach includes the addition
of a new module: africas_talking_sid. This module monitors the session that is created before dispatching the call. By doing this, the session can be found in africas_talking_httpd when receiving a request.
This workaround was done because Africa's Talking doesn't allow an Url param in the post request.