Rework: save session vars only to call_logs
The call log now longer saves the js context but only the session vars,
and only when needed to (otherwise it's an empty proplist). The session
vars are serialized to YAML so we can process them in Ruby as well as
Erlang.

The active GC now reconstructs the callback url from the project
associated to the call log, and authenticates the callback (this was
previously overlooked).

The `js_context` and `callback_url` columns have been dropped from the
`call_logs` table as we don't need them anymore.